Split in multiple files
Refactoring
